+++ src/main/kotlin/io/openfuture/chain/component/converter/transaction/impl/URewardTransactionEntityConverter.kt
@@ -6,6 +6,8 @@ import io.openfuture.chain.crypto.signature.SignatureManager
 import io.openfuture.chain.crypto.util.HashUtils
 import io.openfuture.chain.domain.transaction.BaseTransactionDto
 import io.openfuture.chain.domain.transaction.data.RewardTransactionData
+import io.openfuture.chain.entity.transaction.RewardTransaction
+import io.openfuture.chain.util.TransactionUtils
 import io.openfuture.chain.entity.transaction.unconfirmed.URewardTransaction
 import org.springframework.stereotype.Component
 
@@ -35,7 +37,7 @@ class URewardTransactionEntityConverter(
             data.senderAddress,
             HashUtils.toHexString(keyHolder.getPublicKey()),
             SignatureManager.sign(data.getBytes(), keyHolder.getPrivateKey()),
-            getHash(data, keyHolder.getPublicKey(), keyHolder.getPrivateKey())
+            TransactionUtils.createHash(data, keyHolder.getPublicKey(), keyHolder.getPrivateKey())
         )
 
 }