+++ src/main/kotlin/io/openfuture/chain/component/converter/transaction/impl/UTransferTransactionEntityConverter.kt
@@ -5,6 +5,7 @@ import io.openfuture.chain.domain.rpc.transaction.BaseTransactionRequest
 import io.openfuture.chain.domain.transaction.BaseTransactionDto
 import io.openfuture.chain.domain.transaction.data.TransferTransactionData
 import io.openfuture.chain.entity.transaction.unconfirmed.UTransferTransaction
+import io.openfuture.chain.util.TransactionUtils
 import org.springframework.stereotype.Component
 
 @Component
@@ -31,7 +32,7 @@ class UTransferTransactionEntityConverter : ManualTransactionEntityConverter<UTr
             request.data!!.senderAddress,
             request.senderPublicKey!!,
             request.senderSignature!!,
-            getHash(request)
+            TransactionUtils.createHash(request.data!!, request.senderPublicKey!!, request.senderSignature!!)
         )
 
 }