+++ src/main/kotlin/io/openfuture/chain/component/converter/transaction/impl/UVoteTransactionEntityConverter.kt
@@ -5,6 +5,7 @@ import io.openfuture.chain.domain.rpc.transaction.BaseTransactionRequest
 import io.openfuture.chain.domain.transaction.BaseTransactionDto
 import io.openfuture.chain.domain.transaction.data.VoteTransactionData
 import io.openfuture.chain.entity.transaction.unconfirmed.UVoteTransaction
+import io.openfuture.chain.util.TransactionUtils
 import org.springframework.stereotype.Component
 
 @Component
@@ -33,7 +34,7 @@ class UVoteTransactionEntityConverter : ManualTransactionEntityConverter<UVoteTr
             request.data!!.senderAddress,
             request.senderPublicKey!!,
             request.senderSignature!!,
-            getHash(request),
+            TransactionUtils.createHash(request.data!!, request.senderPublicKey!!, request.senderSignature!!),
             request.data!!.voteType.getId(),
             request.data!!.delegateKey
         )