+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Serialization mechanism to support kotlin classes
+- Structure of transactions
 
 ## [0.3.0] - 2018-07-13
 ### Added